rename slider to pascal case
