+++ src/components/MonthsSlider.css
